+++ src/consensus.rs
@@ -226,6 +226,7 @@ macro_rules! impl_fixed_array {
 }
 
 impl_fixed_array!(6);
+impl_fixed_array!(16);
 impl_fixed_array!(32);
 impl_fixed_array!(33);
 impl_fixed_array!(64);