+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.5.1] - 2022-08-15
+
 ### Added
 
 - Offer `uuid` of type `Uuid` ([#292](https://github.com/farcaster-project/farcaster-core/pull/292))
@@ -149,7 +151,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Messages exchanged between farcaster-node's microservices
 - Tasks and blockchain events used by syncers
 
-[Unreleased]: https://github.com/farcaster-project/farcaster-core/compare/v0.5.0...HEAD
+[Unreleased]: https://github.com/farcaster-project/farcaster-core/compare/v0.5.1...HEAD
+[0.5.1]: https://github.com/farcaster-project/farcaster-core/compare/v0.5.0...v0.5.1
 [0.5.0]: https://github.com/farcaster-project/farcaster-core/compare/v0.4.4...v0.5.0
 [0.4.4]: https://github.com/farcaster-project/farcaster-core/compare/v0.4.3...v0.4.4
 [0.4.3]: https://github.com/farcaster-project/farcaster-core/compare/v0.4.2...v0.4.3